ref(installer): snapper/timeshift setup moves to disk/snapshots.py
* grub-btrfsd override comes along from bootloader/install.py, it only served this
* DiskLayoutConfiguration.btrfs_snapshot_type() replaces the installer helper,
  add_bootloader reads it for the UKI standalone-initramfs decision
* package list moves with it, schema_gen follows
